// clinic/src/main/web/lib/model/clinic.model.ts
export interface Patient {
    id?: number;
    hospitalNum?: string;
    surname?: string;
    otherNames?: string;
    statusAtRegistration?: string;
    dateRegistration?: Moment;
    dateBirth?: Moment;
    gender?: string;
    facility?: Facility;
    uuid?: string;
}
// ...
export interface ChronicCare {
    id?: number;
    uuid?: string;
    patient?: Patient;
    facility?: Facility;
    dateVisit?: Moment;
    currentStatus?: string;
    clinicStage?: string;
    clientType?: string;
    pregnancyStatus?: string;
    lastCd4?: number;
    dateLastCd4?: Moment;
    lastViralLoad?: number;
    dateLastViralLoad?: Moment;
    eligibleCd4?: boolean;
    eligibleViralLoad?: boolean;
    plhivSymtomaticHiv?: boolean;
    plhivAsymtomaticCD4L500?: boolean;
    plhivActiveTb?: boolean;
    plhivPregnantAfter1stTrimester?: boolean;
    plhivL5Years?: boolean;
    ipt?: boolean;
    inh?: boolean;
    tbTreatment?: boolean;
    dateStartedTbTreatment?: Moment;
    tbReferred?: boolean;
    eligibleIpt?: boolean;
    bodyWeight?: number;
    height?: number;
    muac?: number;
    muacPediatrics?: string;
    bmi?: string;
    muacPregnant?: string;
    supplementaryFood?: boolean;
    nutritionalStatusReferred?: boolean;
    sexuallyAbused?: boolean;
    essentialsDeniedByPartner?: boolean;
    sexuallyAbusedReferred?: boolean;
    essentialsDeniedByPartnerReferred?: boolean;
    hypertensive?: boolean;
    firstHypertensive?: boolean;
    bpAbove14090?: boolean;
    bpReferred?: boolean;
    diabetic?: boolean;
    firstDiabetic?: boolean;
    dmReferred?: boolean;
    missedArvs?: string;
    missedArvsServicesProvided?: boolean;
    statusDisclosedToPartner?: boolean;
    partnerStatusKnown?: boolean;
    useCondomsAlways?: boolean;
    useCondomsAlwaysServicesProvided?: boolean;
    opportunisticInfections?: boolean;
    missedCotrim?: string;
    weeklyAlcoholConsumption?: boolean;
    weeklyAlcoholConsumptionServicesProvided?: boolean;
    washServicesProvided?: boolean;
    useInsecticideNets?: boolean;
    cervicalCancerScreening?: boolean;
    activeMemberOfSG?: boolean;
    familyPlanning?: boolean;
    basicCareKits?: boolean;
    disclosureCounseling?: boolean;
    socialServices?: boolean;
    legalServices?: boolean;
    linkageToIGA?: boolean;
    otherServices?: boolean;
    cervicalCancerScreeningWithinPastYear?: boolean;
    cervicalCancerScreeningWithinPastYearReferred?: boolean;
    wantPregnancyWithinAYear?: boolean;
    wantPregnancyWithinAYearReferred?: boolean;
    currentlyUsingContraceptive?: boolean;
    currentlyUsingContraceptiveReferred?: boolean;
    useInsecticideBedNet?: boolean;
    useInsecticideBedNetReferred?: boolean;
    pregnantIntermittentPreventiveTherapy?: boolean;
    pregnantIntermittentPreventiveTherapyReferred?: boolean;
}

export interface DMScreen {
    id?: number;
    description?: string
}

export interface TBScreen {
    id?: number;
    description?: string
}

export interface ChronicCareTB {
    id?: number;
    description?: string;
    tbScreen: TBScreen;
}

export interface ChronicCareDm {
    id?: number;
    description?: string;
    dmScreen: DMScreen;
}

export interface ChronicCareVm {
    chronicCare?: ChronicCare;
    dms?: ChronicCareDm[];
    tbs?: ChronicCareTB[];
}